src/requirements.txt
google-cloud-logging
Flask
google-cloud-firestore
firebase-admin